Readme

+++ scripts/CA/01.enrollTLS.CA.Admin.sh
@@ -2,15 +2,20 @@
 export PATH=$PATH:/Users/cornoro/projects/hyperledger/fabric-samples/bin
 PWD=`pwd`
 
-cp /tmp/hyperledger/tls-ca/crypto/ca-cert.pem /tmp/hyperledger/tls-ca/crypto/tls-ca-cert.pem
+if [ -s /tmp/hyperledger/tls-ca/crypto/ca-cert.pem ]; then
+    cp /tmp/hyperledger/tls-ca/crypto/ca-cert.pem /tmp/hyperledger/tls-ca/crypto/tls-ca-cert.pem
 
-export FABRIC_CA_CLIENT_TLS_CERTFILES=/tmp/hyperledger/tls-ca/crypto/tls-ca-cert.pem
-export FABRIC_CA_CLIENT_HOME=/tmp/hyperledger/tls-ca/admin
-echo "FABRIC_CA_CLIENT_TLS_CERTFILES=$FABRIC_CA_CLIENT_TLS_CERTFILES"
-echo "FABRIC_CA_CLIENT_HOME=$FABRIC_CA_CLIENT_HOME"
-fabric-ca-client enroll -d -u https://tls-ca-admin:tls-ca-adminpw@0.0.0.0:7052
-fabric-ca-client register -d --id.name peer1-org1 --id.secret peer1PW --id.type peer -u https://0.0.0.0:7052
-fabric-ca-client register -d --id.name peer2-org1 --id.secret peer2PW --id.type peer -u https://0.0.0.0:7052
-fabric-ca-client register -d --id.name peer1-org2 --id.secret peer1PW --id.type peer -u https://0.0.0.0:7052
-fabric-ca-client register -d --id.name peer2-org2 --id.secret peer2PW --id.type peer -u https://0.0.0.0:7052
-fabric-ca-client register -d --id.name orderer1-org0 --id.secret ordererPW --id.type orderer -u https://0.0.0.0:7052
+    export FABRIC_CA_CLIENT_TLS_CERTFILES=/tmp/hyperledger/tls-ca/crypto/tls-ca-cert.pem
+    export FABRIC_CA_CLIENT_HOME=/tmp/hyperledger/tls-ca/admin
+    echo "FABRIC_CA_CLIENT_TLS_CERTFILES=$FABRIC_CA_CLIENT_TLS_CERTFILES"
+    echo "FABRIC_CA_CLIENT_HOME=$FABRIC_CA_CLIENT_HOME"
+    fabric-ca-client enroll -d -u https://tls-ca-admin:tls-ca-adminpw@0.0.0.0:7052
+    fabric-ca-client register -d --id.name peer1-org1 --id.secret peer1PW --id.type peer -u https://0.0.0.0:7052
+    fabric-ca-client register -d --id.name peer2-org1 --id.secret peer2PW --id.type peer -u https://0.0.0.0:7052
+    fabric-ca-client register -d --id.name peer1-org2 --id.secret peer1PW --id.type peer -u https://0.0.0.0:7052
+    fabric-ca-client register -d --id.name peer2-org2 --id.secret peer2PW --id.type peer -u https://0.0.0.0:7052
+    fabric-ca-client register -d --id.name orderer1-org0 --id.secret ordererPW --id.type orderer -u https://0.0.0.0:7052
+else
+    echo "CAs docker not running"
+    exit 1
+fi
